Extract the mappers into arsox-harness
The normalization layer is the reason this project exists, and it was only
reachable by depending on the satellite. An application that already schedules
its own containers, supervises its own processes and owns its own storage still
wants the answer to what a harness just did, and getting it meant compiling an
HTTP server, a SQLite store, a job queue and an LLM proxy to reach a pure
function.

For a consumer that runs the harness inside a locked-down container those are
not merely unused dependencies, they are attack surface beside a model running
with tool approval disabled.

harness/mod.rs and harness/claude.rs move unchanged: they were already sans-IO,
which is what made this an extraction rather than a rewrite. The new crate takes
four dependencies and no runtime. runner.rs and spawn.rs stay in the satellite,
which now depends on the crate and re-exports the mapper types so everything
above it reads as before.

The conformance fixture moves with the mapper it proves, which is where it
belongs: it is evidence about Claude's output, not about the satellite.

Adds docs/core-crates.md, which records the split, why arsox-harness must stay
free of I/O, and that publishing goes by tagged git dependency rather than
GitHub Packages, since GitHub Packages has no Cargo registry.
