+++ crates/arsox-harness/fixtures/claude/tool-call.jsonl
